cicd: use CR checker from registry (#852)
* cicd: use CR checker from registry
* cicd: add missing permissions

Decouple the copyright checker and use the module from our custom BZL registry
When you call a reusable workflow (workflow_call) that requires permissions, GitHub mandates that you explicitly grant those permissions from the caller, or the job fails.
This is to prevent a situation where someone could submit a malicious PR that runs code with write access to issues or PRs.

Addresses: #813

Signed-off-by: Dan Calavrezo <dan.calavrezo.ext@qorix.ai>